core/server/pyproject.toml
[tool.poetry]
name = "fl-server"
version = "0.1.0"
description = "Federated Learning aggregation server"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.0.0"
numpy = "^1.24.0"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
